### Sukumu coffee Testing 2.
### Sukumu Style
The name of the coffee is then usually the place it was grown to whatever degree available.

### Sukumu Sustainable farming
What grows where and how it is grown are a matter of choice and careful consideration for nature and communities.

### Sukumu Direct sourcing
Direct trade is a form of sourcing practiced by some coffee roasters. 

### Sukumu profits
We want to truly empower the communities that bring amazing coffee to you.